calendar is fully functioning again!
