+++ css/app.css
@@ -1,1 +1,2 @@
-@import "../node_modules/fullcalendar/dist/fullcalendar.css"
+@import "../node_modules/fullcalendar/dist/fullcalendar.css";
+@import "../node_modules/bootstrap/dist/css/bootstrap.css"
